[refactor] added chaincfg to transaction
