+++ compat/bip32/extendedkey.go
@@ -19,10 +19,10 @@ import (
 	"math/big"
 	"sync"
 
-	"github.com/bitcoin-sv/go-sdk/chaincfg"
 	base58 "github.com/bitcoin-sv/go-sdk/compat/base58"
 	ec "github.com/bitcoin-sv/go-sdk/primitives/ec"
 	crypto "github.com/bitcoin-sv/go-sdk/primitives/hash"
+	chaincfg "github.com/bitcoin-sv/go-sdk/transaction/chaincfg"
 )
 
 const (
