+++ compat/wif/wif_test.go
@@ -7,8 +7,8 @@ package compat
 import (
 	"testing"
 
-	"github.com/bitcoin-sv/go-sdk/chaincfg"
 	ec "github.com/bitcoin-sv/go-sdk/primitives/ec"
+	chaincfg "github.com/bitcoin-sv/go-sdk/transaction/chaincfg"
 )
 
 func TestEncodeDecodeWIF(t *testing.T) {
